refactor(docs): extract shared toolkit schema processing into lib/toolkit-schema
Deduplicates processParams/processApiSchema/toolFromApi that were
duplicated across the server page and client toolkit-detail component.
